Mean added

+++ RocketApp/app/src/main/java/com/example/rocketapp/TrialFragment.java
@@ -84,7 +84,7 @@ public class TrialFragment extends DialogFragment {
                             break;
                         case MeasurementTrial.TYPE:
                             if (Validate.intInRange(inputEditText, 0, Integer.MAX_VALUE, true)) {
-                                callback.callBack(new MeasurementTrial(Integer.parseInt(inputEditText.getText().toString())));
+                                callback.callBack(new MeasurementTrial(Float.parseFloat(inputEditText.getText().toString())));
                                 alertDialog.dismiss();
                             }
                             break;
